+++ AimsProject/Aims.java
@@ -1,4 +1,4 @@
-package Lab_2;
+package AimsProject;
 
 public class Aims {
     public static void main(String[] args) {
