+++ AimsProject/Cart.java
@@ -1,4 +1,4 @@
-package Lab_2;
+package AimsProject;
 
 public class Cart {
     public static final int MAX_NUMBER_ORDERED = 20;
