+++ OtherProjects/Lab_1/Exercies/InputFromKeyboard.java
