+++ OtherProjects/Lab_1/Exercies/PrintingStar.java
