+++ OtherProjects/Lab_1/SolveEquation.java
